Add `CIME.namelist.write`.
This namelist-writing function is the counterpart to
`CIME.namelist.parse`. This is not yet hooked up to anything, but it was
tested by reading in a CAM namelist, writing the namelist back out, and
then reading the result back in, to ensure that a `write` followed by a
`parse` will give back identical data to the original object in memory.