Add `groupless` mode to `CIME.namelist`.
For processing of `user_nl_*` files, we don't want to require users to
know what the group name associated with each variable is. This commit
therefore adds the ability to ignore group names when parsing namelist
input, as well as to parse namelist-like input where no group names are
present at all.