Allow namelist variables to be redefined.
If a single namelist variable is defined more than once, that's allowed.
Typically the new value overwrites the old one, but if there are null
values in the new setting, then the old value wins. This change allows
namelists such as the following:

```
foo_list = 8*'I',
foo_list = , , 'A'
```

This is equivalent to:

```
foo_list = 'I', 'I', 'A', 'I', 'I', 'I', 'I', 'I'
```