Add `NamelistDefaults` attribute matching.
This is largely the same as the Perl version: attributes that don't
match the value for the current case cause a default to be discarded,
and attributes that do match increase the score used to pick a default.
In case of tie, the first matching value wins, so that less specific
values followed by more specific values work as expected.

One difference from the Perl version is that attributes that are not
defined in the current case, but *are* defined in the file, also cause a
default to be discarded. This is to deal with cases like this:

```
<foo opt1='a' unusual_option='bar'>val1</foo>
<foo opt2='b'>val2</foo>
```

For a case with opt1 = 'a' and opt2 = 'b', we probably want the latter
value to be chosen. (However, this should rarely come up, since most
options are always defined in a given case, even if just to some null
value such as "none", "0", "false", or "off".)