Adjust `NamelistDefaults` attribute matching.
This contains one bug fix: if defaults are found for a variable, but
none of them match the requested configuration, `get_value` now returns
`None` instead of raising an exception.

There is also a new feature. Configuration information can now be passed
to `get_value` in the `attribute` argument, and will be used together
with the `attributes` used to construct a `NamelistDefaults` object when
selecting a default.