Using latest Cray env variable for dynamic linking
Using CRAYPE_LINK_TYPE instead of XTPE_LINK_TYPE to indicate
dynamic linking when building ACME on Titan with the intel
compiler.

Without this fix cmake adds the "-rdynamic" flag to the linker
flags and aborts with the following error message,

"/usr/bin/ld: dynamic STT_GNU_IFUNC symbol strcmp' with pointer equality
in/usr/lib64/gcc/x86_64-suse-linux/4.3/../../../../lib64/libc.a(strcmp.o)'
can not be used when making an executable; recompile with -fPIE and
relink with -pi"

This issue did not occur with cmake 2.8.11.2 . However it occurs with
older (cmake 2.8.10.2 - the default on titan) and newer (cmake 3.2.3)
versions.

Fixes #995

[BFB]