Adds support for next atmospheric compset (FC5AV1C-05)
The PR adds support for next atmospheric compset which is in works at
this point. A few namelist parameters are added which will assist in
the tuning of FC5AV1C-05 compset.

Following is the list of namelist parameters:
1. cld_sed   : scale factor for cloud sedimentation velocity
2. clubb_tk1 and clubb_tk2 : temperature thresholds determining the
fraction of condensate detrained from ZM as ice
3. clubb_wpxp_L_thresh : Lscale threshold: damp C6 & C7  [m]
4. gust_fac : wind gustiness factor

All these changes are BFB except for clubb_tk1 and clubb_tk2 change.
This changes answers but it should be at the round-off level

[non-BFB]
[NML]