Not resuing file created on a diff iosys
This change makes sure that we open the file on the new iosys
(that sets the rearr opts) rather than re-using the file and var
that were opened on the default iosystem. Although the previous
code worked because we are essentially operating on a dup of
the default comm, it is not a mode supported by pio.

* All common names are exported to an internal module
* A new function that just opens and inquires the var is used now
  to re-open the file