Switch to choose btw irsend and isends
Adding a compile-time switch to switch between MPI_Irsends and
MPI_Isends. With PIO1 we have seen software stacks where due to
bugs in MPI_Irsends the code hangs (the same code runs fine with
isends on the same machine and with irsends on other machines).
The default is still mpi_irsend.