Reset rearr opts to default if user does not
This commit makes sure that that rearranger options are valid,
reset to defaults if the user does not set them. Also setting
max pend req to 0 if the user chooses to use a collective (the
current code uses max pend req = 0 as the cue to use collectives
for rearranging data).

The new defaults disable handshake, isend and sets the max
pending reqs to unlimited if the user disables flow control in
that direction.