Revert pio1 cmake path to point into pio2
Add back in the pointing of pio1 into pio2/cmake, for now.

The CMakelists.txt file for pio1/pio used to point into pio2/cmake for functions like FindNetCDF.cmake.
commit "Dont point to pio2 cmake modules by default"
PR#1352

This was changed in pio1/pio to point to "${CMAKE_CURRENT_SOURCE_DIR}/cmake" which doesn't exist.

Without a FindNetCDF, the ACMCE build of the HOMME test to fail. 

Changing this back to HOMME build works again until a better fix is found or we switch to pio2.