Changing jobmin to 1 for batch q on blues
The jobmin for the "batch" queue on blues was set to 64 before
and hence jobs that required more than 1 hr (> walltimemax for
shared/debug queue) and less than 64 procs (jobmin for batch
queue on blues is 64) did not fit in any available queue.

Setting jobmin to 1 for blues+batch queue so that longer jobs
that don't fit in the shared/debug queue fit the batch queue.

Fixes ACME-Climate/ACME#1484