Fix indices and some dbg statements
Fix the loop indices to correctly parse the list of decomp files.
Note: $#array gives the last index not the size of the array.

By default assume that two files of same size are dups and need
removal and set it to false if any line differs btw the files.

Also adding some debug statements.

Now duplicate files are correctly deleted by the script