Check stack trace before comparing lines
Now checking the beginning of stack trace before comparing the
lines in the files.

Without this fix we might never be able to detect the beginning
of some stack traces("Obtained ..." line) because that line could
be identical for dup files and we continue comparing the next
line. Previously we only checked for the beginning of stack trace
when the lines did't match.