Make sure to match stack trace on both files
Make sure that we declare files are duplicate when lines being
compared on both (not just the first) files have the beginning
of a stack trace.

Although unlikely but we don't want to declare two files of the
same size as dups when we hit the beginning of the stack trace
on the first file (without checking if we are at the beginning
of the stack trace on the second file).