Use HIST_OPTION=never for LVIRTUALGLC test
Need HIST_OPTION=never for this test because we expect answer changes in
the l2x fields that are separated by elevation class. So for this test,
we just ensure that there are no answer changes in gridcell-average CLM
fields, via comparison of the CLM history files. (If we had a way to
exclude individual fields from the list of compared fields, then we
could compare cpl hist files, excluding certain l2x fields.)