Change use of CIME dir to an absolute path
The relative path used to define where CIME is located can be dangerous
in certain places. The way it was being used was fine, but to add an external
project for pio was going to make it a bad idea since the path would be
interpreted from a different context. So switching to the path to CIME to be
absolute will make it crystal clear where CIME will be in all contexts.

The input to the cmake build is still a relative path so host codes won't
need to worry about this detail or getting it right. They will still give
the location based on their context.