Partial fix for incorrect type of 'mpi_type' in pioc_support.c
The type for mpi_type in the routines inq_file_metadata and PIOc_openfile_retry
in pioc_support.c is int but should be MPI_Datatype because mpi_type is used in
calls to, e.g., MPI_Type_size (https://www.open-mpi.org/doc/v2.0/man3/MPI_Type_size.3.php).
In most implementations of MPI, the size of MPI_Datatype is the same as the size
of int, but in some versions of OpenMPI, MPI_Datatype actually has a size of 8.
This leads to memory errors and segfaults in PIO.