Add and clarify pieces that would have been helpful when I was trying to add a test to the prebeta testing.
Expand the introduction.

Expand the test name syntax description,
using the same table format as the TESTTYPE table,
instead of a bulleted list, for consistency.

Add a more complete table of MODIFIERS of TESTTYPE,
replacing a brief text list.

Add a table of GROUP-TESTMODS, for completeness and clarity.

Expand description of create_test and query_testlists.

Add a CESM subsection to the "Adding tests" section,
since the E3SM instructions don't apply to CESM.