Report BFAIL for missing baseline files if there were no baseline fails
THIS IS UNTESTED: I haven't done manual tests of this or added unit
tests for this yet.

This implements the logic that used to be in
component_compgen_baseline.sh (as I note in
https://github.com/ESMCI/cime/pull/2744#issuecomment-412208219):

1. The given baseline directory exists, but there are no history files
   there. (What I'd like: BFAIL on the BASELINE line.)

2. The given baseline directory exists. There are some history files
   there, but not all that there should be. The history files there are
   bit-for-bit with the new test. (What I'd like: BFAIL on the BASELINE
   line.)

3. The given baseline directory exists. There are some history files
   there, but not all that there should be. At least one history file is
   _not_ bit-for-bit with the new test. (What I'd like: FAIL, but _no_
   BFAIL on the BASELINE line.)

I also cleaned up the creation of the test status line for BASELINE
failures, making this the responsibility of compare_baseline (in
hist_utils.py) rather than the previous checking for whether "\n" was in
the full comments. (This removes some duplication between
system_tests_common.py and compare_test_results.py.)