Allow specific tests to ignore field list differences
Due to recent changes to cprnc, differences in field lists (i.e., if one
case has some diagnostic fields that are missing from the other case)
are now treated as differences. However, this is a problem for some new
tests that are wanted, such as ensuring that turning on carbon isotopes
doesn't change answers.

This commit allows a particular test type to dictate that field list
differences should be ignored. For tests that use SystemTestsCompareTwo,
this is as simple as adding an argument to the SystemTestsCompareTwo
constructor. For other tests, a new argument can be added to the call to
self._component_compare_test. (Differences in the values in shared
fields are still treated as differences.)