Pass COMP_LND to build
All of the COMP_* variables are passed to the build via environment
variables when doing a build, but not when doing a clean build. COMP_LND
is needed for a clean build as well, because the location of the land
build depends on the value of COMP_LND (CLM uses the shared build
area). Passing it via _CMD_ARGS_FOR_BUILD ensures that it is provided in
the clean build as well as standard build situation.

Resolves ESCMI/cime#3286