Remove sphinx_rtd_theme from list of extensions
When it was in the list of extensions, I got this warning when building
documentation:

WARNING: extension 'sphinx_rtd_theme' has no setup() function; is it really a Sphinx extension module?

I'm thinking that we're using an older version of sphinx_rtd_theme that
doesn't support this. From the documentation, I see:

   Adding this theme as an extension is what enables localization of
   theme strings in your translated output. If these strings are not
   translated in your output, either we lack the localized strings for
   your locale, or you are using an old version of the theme.