When writing .env_mach_specific files, only write settings for main job
Previously, the .env_mach_specific.sh and .env_mach_specific.csh files
had settings for all jobs. At least on cheyenne, this led to the
settings for the st_archive job being written after the settings for the
main job - in particular, setting MPI_USE_ARRAY=false. I typically want
the settings for the main job. This commit changes the behavior so that
the env_mach_specific files just contain the settings for the main job.

Note that this doesn't change the behavior when running the standalone
configure tool, because I'm not sure how to fix that case.