Allow 0 max pending I/O requests
Allow user to specify the number of maximum pending I/O requests
as 0.

The number of maximum pending I/O requests by default in
env_run.xml is now set to -2 (not 0). A setting of -2 indicates
that CIME calculates the value of max pending I/O requests
internally

Also see PR #3499 that required this behavior.