Avoid building CDEPS with CTSM's LILAC
For now, avoid building CDEPS with CTSM's LILAC because it's not needed
in this configuration and CDEPS relies on unreleased ESMF
code. Eventually we will require CDEPS (for its streams functionality),
at which point CDEPS should only require released ESMF code; then we
should revert the changes in this commit.