More build updates for e3sm
The only impacts to other models is a slight change to the timing
output at the end of case_build. I don't think it's necessary to continue
to output 'Time spent not building' since we've knocked that way
down and it's only a tiny fraction of build time now. Also, the other
timings printing in units 'seconds' so the final print should use that
too instead of 'sec'.