Avoid adding CDEPS things to inc & lib paths with LILAC
This is now needed with https://github.com/ESMCI/cime/pull/3770, given
some other changes to the build that have been made in the last few
months (I think the removal of the conditional in the Makefile in
https://github.com/ESMCI/cime/pull/3779, but I'm not sure).

This and the changes in #3770 will be backed out once we require pieces
of CDEPS in the LILAC build (specifically for the streams
functionality).

Note that this is a cherry-pick of 0dbcf4971.