Add ATM_FLUX_INTEGRATION_METHOD option.
This commit adds an env_run.xml variable that controls whether the new
implicit stress calculations will be used. This also provides a single
code that can use either the explicit or implicit stress calculation.

The explicit (default) setting is bit-for-bit with the version of E3SM
master that this code (and associated E3SM commit) was branched from.