Reset a test that is rerun after a run failure
When running create_test with the --retry option set to 1 or more: If a
test (ERS, ERI and possibly others) failed in the last run, some xml
values would be set incorrectly in the retried run. This is because
--retry sets use_existing, which in turn changes a FAILed RUN phase into
a PEND. Then, in the call to _resetup_case(RUN_PHASE) from
system_tests_common:run, the test reset wasn't done, since PENDing
phases are not reset.

This commit solves this problem by adding a comment that the now-PENDing
phase is actually being RERUN, and checking for that in _resetup_case.

There may be other situations where we should add this RERUN comment,
but this change fixes my original issue, and it wasn't obvious to me
what, if anything, else should be changed from looking through the
code (especially test_scheduler.py).

Resolves ESMCI/cime#3933