Rework mechanism for avoiding error in unit tests & doc tests
Reasons I prefer this method:

(1) (Main reason) It works for python -m unittest discover (the previous
method did not)

(2) I found the other method hard to understand

(3) It makes me nervous to have unit test-specific code sprinkled in the
production code, especially when that test-specific code is hard to
understand: it's hard to be confident that things will work right in
production.

For the doctests, it looks like they were creating a case in the current
directory (wherever that happens to be), which feels like it could cause
trouble in some circumstances, so it felt safer to fix these tests
rather than using a workaround that allowed them to try to create a case
in the current directory.