cime_bisect: Improve robustness when using custom script
We do not want to check for --wait if user is using a custom script.
The check for --wait required knowing if machine is a batch system
which required Machines which opened up a lot of potential failures.

cime_bisect should now work, for custom scripts, even on machines
on which cime is not supported.