Add optional config_archive element: hist_file_ext_regex
This is needed for CISM, where we have history file names with a
two-part extension: instead of casename.cism.h.*, we have
casename.cism.gris.h.*, casename.cism.ais.h.*, etc.

Here is the explanation copied from a comment in config_archive.xsd:

hist_file_ext_regex is used in a specific context: This specifies the
regex capture group following the component name that is used to
differentiate between different history file streams for a given
component. In most cases, this can be excluded, falling back on the
general-purpose extension regex, "\w+". However, this is needed in
situations where the extension is more complex, such as when there is a
multi-part extension with embedded '.' characters. For example, the CISM
ice sheet model uses a hist_file_ext_regex of "\w+\.\w+" because its
history file names have a two-part extension, with the first part giving
the ice sheet name (e.g., gris.h, ais.h).