Improve Macro.make generation
I got a bug report from someone for whom the original implementation
did not work because he had his grep aliased to produce colorful
output, which was then caused lots of weird characters to get added
to his Macro.make file.

This PR encapsulates this capability and converts more of the implementation
to pure python to avoid portability issues like this.