Fix machine check for some unusual cases
This fixes the machine check in two situations, both of which come up
when building CTSM via the LILAC build process for a user-defined
machine:

(1) If the machine name can't be probed - e.g., if it was set explicitly
in the create_newcase command because it can't be probed on this machine
- then you need to specify a machine name explicitly in order to
initialize the Machines object.

(2) If the information for this machine is specified in an
`extra_machines_dir`, then this directory needs to be provided when
initializing the Machines object (otherwise you get an error like
`ERROR: No machine YOUR_MACHINE_NAME found`).