Changes needed for E3SM
Our PIO (scorpio) does not work with the latest genf90 changes,
specifically the addition of 'short' support. This commit changes
the gen90 submodule to point to a genf90 commit with short support
reverted.

Also, our Jenkins jobs rely on the 'reuse host credential' feature
and so we must have all submodules be using the ssh protocol.