Check number of hist file comparisons in system_test_common
Every file that gets copied by copy_histfiles should later be
compared by compare_test. We ran into a situation where the
hist_file_extension for a component worked when suffix= but
not when suffix!=. This caused hist files for that component to
get silently skipped in various system tests that do internal compares.

This PR will make the scenario above be reported as a failed comparison
so it can't go unnoticed.