case.submit: Always try to download input data
We don't want test cases downloading data while on a compute node
if it can be avoided. It's harmless to call check_all_input_data
twice, so let's have the case.submit call handle the 99% of cases
where the main case will download all the inputs needed for the
case/test.