Fix bug in path to CAM script cam.case_setup.py
Just prior to the cime6.0.217 tag there was a change in the original code
to call cam.case_setup.py during case setup. The change added a
conditional for file path existence, but the file path added was incorrect.
This causes the script to never be called and GEOS-Chem compsets will
subsequently all fail.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>