Implement dependency tracking support for Fortran submodules
The submodule feature was added in the Fortran 2008 language standard.
A submodule is a program unit that extends a module or another submodule.
This means that submodules can be nested. As such, compiling a submodule
requires that all of its ancestors be compiled first.

This PR enables the `mkDepends` tool to recognize the Fortran `submodule`
statement, and to correctly generate build dependencies for it.