When calculating stop_n in coupling intervals, maintain original, too
This way, in prints of stop_n, we can print the original - i.e., in the
original units. This is particularly important for the logger message,
`doing an {0} {1} initial test with restart file at {2} {1}`, which was
previously giving the wrong stop_n value (it was giving stop_n in
coupling intervals rather than the original value).

Also move some error checks on stop_n to before the use of stop_n to
compute rest_n.