Fix error check in _set_restart_interval
The previous error check for stop_n > 2 was being done on the stop_n
value that was expressed in terms of coupling intervals; this is not the
correct check and allows too-small stop_n values. Rather than trying to
determine the correct check on the stop_n value (which is complicated
now), instead check the end result: the real point of this error check
seems to be to make sure that the final rest_n is between 0 and stop_n
(exclusive), so check that directly.