Allow namelist variables that vary from what's in namelist_definition
Previously, any namelist variable needed to appear in the
namelist_definition xml file. With the changes here, the
namelist_definition file can contain a generic variable name that gives
metadata for one or more final namelist variables with different names.
This is done by adding multiple copies of a namelist_definition entry,
one for each final name.

The use case for this is having configuration variables whose names are
not known ahead of time, but instead are determined at build-namelist
time based on some xml variables. Specifically, I'm planning to use this
to support the addition of arbitrary water tracers, each of which will
have one or more configuration variables whose name contains the name of
that water tracer, where the list of water tracer names is set in an xml
variable. The type and default value will be the same for each water
tracer, so we can take this metadata from a single, generic entry in the
namelist_definition file, but then we need to support an add_default
call for each individual, dynamically-named variable.