Fix tests in TestUnitBaselinesPerformance on my Mac
The problem was that read_baseline_file returns without doing anything
if not os.path.exists(baseline_file). So we need to make it think that
the file exists. Previously one of the failing tests had been doing this
via `os.mknod("/tmp/cpl-mem.log")`, but that was itself giving an error
on my Mac.