Fix handling of namelist variables with uppercase characters
Prior to this commit, variables with uppercase characters in the
namelist definition file could not be modified via user_nl files
(regardless of whether the entry in the user_nl file matched the correct
case or was all lowercase).

This commit fixes this behavior so that entries in the user_nl file are
case-insensitive, while preserving the case as defined in the namelist
definition file in the final output.