Fix test_unit_xml_tests for new case.get_value call
The previous commit introduced an extra call to case.get_value to get
SYSTEM_TESTS_DIR for component=allactive. This broke the mock get_value
returns. To make these returns more robust, I have changed the way
they're done to be based on a dictionary lookup rather than an ordered
list.

This kind of replacement should probably be done for other
case.get_value mocks. Note that I identified the argument values that
needed to be handled with this code:

   for i, call in enumerate(case.get_value.call_args_list):
       print(f"  Call {i}: {call}")