Don't ignore externals directory
There is a valid subdirectory named 'components':
CIME/non_py/src/components. We don't want to ignore that. I thought
about making the gitignore line more specific, but I'm not sure what
we're actually trying to ignore here, so for now I'm just removing it.
And while I'm at it, I'm removing the ignore for some other things that
I don't think need to be here, or are too general.